QComp

Documentation for QComp.

QComp.Data

QComp.GenericParams

QComp.InvariantParams

QComp.Params

QComp.QSVT_square

QComp.Usp

QComp.block_encode2

QComp.build_QCNN

QComp.conv_Ry

QComp.eval_Usp

QComp.eval_full_grad

QComp.eval_full_loss

QComp.eval_grad

QComp.eval_grad

QComp.eval_loss

QComp.eval_loss

QComp.expand_params

QComp.initialize_params

QComp.loss

QComp.pcp

QComp.reduce_params

QComp.sigmoid

QComp.test_model

QComp.train_test_model

QComp.Data — Type

struct Data{S <: Vector{SS} where SS <: ArrayReg, L <: Vector{LL} where LL <: Real}

Structure representing input data for the model.

Fields

s::S: Input states.

l::L: Labels.

QComp.GenericParams — Type

Base.@kwdef mutable struct GenericParams{NN<:Integer, CC<:ChainBlock, TT<:AbstractVector{T} where T<:Real} <:Params

Defines parameters for a quantum circuit without translational invariance.

Fields

n::NN: Dimension of the quantum register (must be a power of 2).

circ::CC: Circuit structure of the QCNN.

params::TT = Float64[]: Vector of parameters, initialized to an empty vector of Float64 values.

QComp.InvariantParams — Type

Base.@kwdef mutable struct InvariantParams{NN<:Integer, CC<:ChainBlock, TT<:AbstractVector{T} where T<:Real} <:Params

Defines parameters for a quantum circuit with translational invariance.

Fields

n::NN: Dimension of the quantum register (must be a power of 2).

circ::CC: Circuit structure of the QCNN.

params::TT = Float64[]: Vector of parameters, initialized to an empty vector of Float64 values.

QComp.Params — Type

abstract type Params end

Abstract type representing parameters for the QCNN.

QComp.QSVT_square — Method

QSVT_square(n, d, phis, A)

Build the QSVT (Quantum Singular Value Transform) sequence.

Arguments

n: Number of qubits.

d: Dimensionality parameter.

phis: Array of phase angles.

A: Matrix to transform.

Returns

The QSVT sequence circuit.

QComp.Usp — Method

Usp(phis, a)

Build the signal processing operator.

Arguments

phis: Array of phase angles.

a: Angle parameter.

Returns

The signal processing operator circuit.

QComp.block_encode2 — Method

block_encode2(n, A)

Encode a block using a method.

Arguments

n: Number of qubits.

A: Matrix to encode.

Returns

The encoded block circuit.

QComp.build_QCNN — Method

build_QCNN(n)

Build a quantum circuit for a Quantum Convolutional Neural Network (QCNN).

Arguments

n: Dimension of the quantum register.

Returns

A quantum circuit representing the QCNN.

QComp.conv_Ry — Method

conv_Ry(n, i, j)

Ansatz for the convolutional layer. Use Ry gate to keep ansatz real (adapted from arxiv:2108.00661v2).

Arguments

n: Dimension of the quantum register.

i: Index of the first qubit.

j: Index of the second qubit.

Returns

A quantum circuit implementing convolutional layers with Ry gates.

QComp.eval_Usp — Method

eval_Usp(x, phis)

Evaluate the signal processing operator.

Arguments

x: Input value.

phis: Array of phase angles.

Returns

The result of applying the signal processing operator.

QComp.eval_full_grad — Method

eval_full_grad(d::Data, p::InvariantParams, sig)

Compute the full gradient of the cost function for input data with parameters with translational invariance.

Arguments

d::Data: Input data and labels.

p::InvariantParams: Model parameters.

sig: Boolean indicating whether to use sigmoid activation.

Returns

The total gradient of the cost function.

QComp.eval_full_loss — Method

eval_full_loss(d::Data, p::Params, sig)

Evaluate the full mean squared error (MSE) loss of a model on input data.

Arguments

d::Data: Input data and labels.

p::Params: Model parameters.

sig: Boolean indicating whether to use sigmoid activation.

Returns

The total mean squared error loss.

QComp.eval_grad — Method

eval_grad(state, p::GenericParams)

Compute the gradient of the cost function for a given state.

Arguments

state: Input state.

p::GenericParams: Model parameters.

Returns

The gradient of the cost function.

QComp.eval_grad — Method

eval_grad(state, p::InvariantParams)

Compute the gradient of the cost function for a given state with parameters with translational invariance.

Arguments

state: Input state.

p::InvariantParams: Model parameters.

Returns

The gradient of the cost function.

QComp.eval_loss — Method

eval_loss(state, p::GenericParams)

Evaluate the loss of a model (with no invariance) applied to a state.

Arguments

state: Input state.

p::GenericParams: Model parameters.

Returns

The loss value.

QComp.eval_loss — Method

eval_loss(state, p::InvariantParams)

Evaluate the loss of a translational invariant model applied to a state.

Arguments

state: Input state.

p::InvariantParams: Model parameters.

Returns

The loss value.

QComp.expand_params — Method

expand_params(p::InvariantParams)

Expand unique translational invariant parameters to a full vector.

Arguments

p::InvariantParams: Parameters object containing circuit information.

Returns

A vector representing expanded parameters.

QComp.initialize_params — Method

initialize_params(p::InvariantParams)

Initialize parameters for a quantum circuit with translational invariance.

Arguments

p::InvariantParams: Parameters object containing circuit information.

Errors

Throws an error if the register dimension is not a power of 2.

QComp.loss — Method

loss(target, xs, phis)

Compute the loss function for a given target, inputs, and phase angles.

Arguments

target: Target value.

xs: Array of input values.

phis: Array of phase angles.

Returns

The loss value.

QComp.pcp — Method

pcp(n, phi)

Generate a projected controlled phase gate.

Arguments

n: Number of qubits.

phi: Phase angle.

Returns

The projected controlled phase gate circuit.

QComp.reduce_params — Method

reduce_params(n, params)

Reduce a vector of parameters to the unique set of parameters in the translational invariant QCNN. Performs the sum of the corresponding parameters in the circuit. Used in evalfullgrad.

Arguments

n: Dimension of the quantum register.

params: Vector of "extended" parameters.

Returns

A vector of unique parameters.

QComp.sigmoid — Method

sigmoid(x)

Compute the sigmoid function.

Arguments

x: Input value.

Returns

The sigmoid of the input value.

QComp.test_model — Method

test_model(d::Data, p::Params, sig)

Evaluate classification accuracy of the model.

Arguments

d::Data: Input data and labels.

p::Params: Model parameters.

sig: Boolean indicating whether to use sigmoid activation.

Returns

preds: Predicted values.

suc_rate: Success rate.

suc_inds: Indices of successful predictions.

QComp.train_test_model — Method

train_test_model(d1::Data, d2::Data, p::Params, iters, sig, lr; output=false)

Train and test the model, outputting final predictions and train/test accuracy.

Arguments

d1::Data: Training data and labels.

d2::Data: Testing data and labels.

p::Params: Initial model parameters.

iters: Number of training iterations.

sig: Boolean indicating whether to use sigmoid activation.

lr: Learning rate for optimization.

output: Boolean indicating whether to output intermediate training information.

Returns

p.params: Final model parameters.

loss_track: Track of loss values during training.

tr_track: Track of training accuracy during training.

te_track: Track of testing accuracy during training.

tr_preds: Predictions on training data.

te_preds: Predictions on testing data.