Graph documentation

Structure

QuantumPrimer.AbstractGraph — Type

AbstractGraph

An abstract type for representing graphs.

QuantumPrimer.Graph — Type

Graph{T<:Vector{Vector{TT}} where TT<:Integer}

A mutable struct representing a graph.

Fields

l::T: A vector of vectors representing the adjacency list of the graph.

Constructor

Graph(l): Creates a Graph instance with the given adjacency list l.

QuantumPrimer.SignedGraph — Type

SignedGraph{T<:Vector{Vector{T}} where T<:Integer, TT<:Vector{TT} where TT<:Integer}

A mutable struct representing a signed graph.

Fields

l::T: A vector of vectors representing the adjacency list of the graph.

s::TT: A vector of integers representing the signs of the edges.

Constructor

SignedGraph(l, s): Creates a SignedGraph instance with the given adjacency list l and edge signs s.

QuantumPrimer.sort_lexicographic — Function

sort_lexicographic(vec::Vector{Vector{Int}}) -> Vector{Vector{Int}}

Sorts a vector of vectors lexicographically and removes duplicates.

Arguments

vec::Vector{Vector{Int}}: The vector of vectors to be sorted.

Returns

Vector{Vector{Int}}: The sorted and unique vector of vectors.

sort_lexicographic(vec_l::Vector{Vector{Int}}, vec_s::Vector{Int}) -> Tuple{Vector{Vector{Int}}, Vector{Int}}

Sorts a vector of vectors lexicographically and a corresponding vector of integers.

Arguments

vec_l::Vector{Vector{Int}}: The vector of vectors to be sorted.

vec_s::Vector{Int}: The vector of integers to be sorted according to vec_l.

Returns

Tuple{Vector{Vector{Int}}, Vector{Int}}: The sorted vectors.

sort_lexicographic(arr::Array) -> Array

Sorts an array of arrays lexicographically.

Arguments

arr::Array: The array of arrays to be sorted.

Returns

Array: The sorted array of arrays.

Line graph

QuantumPrimer.create_line_graph — Function

create_line_graph(G::AbstractGraph) -> Graph

Creates the line graph of a given graph.

Arguments

G::AbstractGraph: The input graph.

Returns

Graph: The line graph of the input graph.

Throws

An error if the input graph has no edges.

create_line_graph(G::SignedGraph) -> SignedGraph

Creates the line graph of a given signed graph.

Arguments

G::SignedGraph: The input signed graph.

Returns

SignedGraph: The line graph of the input signed graph.

Throws

An error if the input graph has no edges.

Mapping to Hamiltonian thermal state

QuantumPrimer.create_interaction_matrix — Function

create_interaction_matrix(G::SignedGraph) -> Matrix{Int}

Creates the interaction matrix of a given signed graph.

Arguments

G::SignedGraph: The input signed graph.

Returns

Matrix{Int}: The interaction matrix of the input signed graph.

QuantumPrimer.create_hamiltonian_matrix — Function

create_hamiltonian_matrix(A::Array{Int, 2}) -> Matrix{ComplexF64}

Creates the Hamiltonian matrix for a given interaction matrix.

Arguments

A::Array{Int, 2}: The interaction matrix.

Returns

Matrix{ComplexF64}: The Hamiltonian matrix.

QuantumPrimer.from_string_to_matrix — Function

from_string_to_matrix(vec::AbstractArray) -> Matrix{ComplexF64}

Converts a vector of matrices (developed initially for Paulis) to a single matrix using the Kronecker product.

Arguments

vec::AbstractArray: The vector of matrices.

Returns

Matrix{ComplexF64}: The resulting matrix.

QuantumPrimer.ρ_gibbs — Function

ρ_gibbs(β::Real, matrix::AbstractMatrix) -> Matrix{ComplexF64}

Computes the Gibbs state for a given Hamiltonian matrix at inverse temperature β.

Arguments

β::Real: The inverse temperature.

matrix::AbstractMatrix: The Hamiltonian matrix.

Returns

Matrix{ComplexF64}: The Gibbs state.

QuantumPrimer.create_purified_state — Function

create_purified_state(β::Real, G::SignedGraph) -> ArrayReg

Creates a purified quantum state for a given signed graph at inverse temperature β.

Arguments

β::Real: The inverse temperature.

G::SignedGraph: The input signed graph.

Returns

ArrayReg: The purified quantum state.

Classification functions

QuantumPrimer.find_negative_eigs — Function

find_negative_eigs(matrix::AbstractMatrix) -> Int

Finds the number of negative eigenvalues of a given matrix.

Arguments

matrix::AbstractMatrix: The input matrix.

Returns

Int: The number of negative eigenvalues.

find_negative_eigs(G::SignedGraph) -> Int

Finds the number of negative eigenvalues of the Hamiltonian matrix of a given signed graph.

Arguments

G::SignedGraph: The input signed graph.

Returns

Int: The number of negative eigenvalues.

QuantumPrimer.find_degeneracy_gs — Function

find_degeneracy_gs(matrix::AbstractMatrix) -> Int

Finds the degeneracy of the ground state of a given matrix.

Arguments

matrix::AbstractMatrix: The input matrix.

Returns

Int: The degeneracy of the ground state.

find_degeneracy_gs(G::SignedGraph) -> Int

Finds the degeneracy of the ground state of the Hamiltonian matrix of a given signed graph.

Arguments

G::SignedGraph: The input signed graph.

Returns

Int: The degeneracy of the ground state.

QuantumPrimer.classify_state_degeneracy — Function

classify_state_degeneracy(n::Int, label::Int; max_iter::Int=1000, use_line_graph::Bool=false) -> SignedGraph

Classifies a random signed graph based on the degeneracy of its ground state.

Arguments

n::Int: The number of vertices.

label::Int: The label indicating the desired degeneracy classification (1 for high degeneracy, -1 for low degeneracy).

max_iter::Int: The maximum number of iterations to attempt, default is 1000.

use_line_graph::Bool: Whether to use the line graph of the random signed graph, default is false.

Returns

SignedGraph: The classified random signed graph.

Throws

An error if a suitable graph cannot be generated within the maximum number of iterations.

QuantumPrimer.classify_state_negative_spectrum — Function

classify_state_negative_spectrum(n::Int, label::Int; max_iter::Int=1000, use_line_graph::Bool=false) -> SignedGraph

Classifies a random signed graph based on the number of negative eigenvalues in its Hamiltonian matrix.

Arguments

n::Int: The number of vertices.

label::Int: The label indicating the desired classification (1 for more than 3 negative eigenvalues, -1 for 3 or fewer negative eigenvalues).

max_iter::Int: The maximum number of iterations to attempt, default is 1000.

use_line_graph::Bool: Whether to use the line graph of the random signed graph, default is false.

Returns

SignedGraph: The classified random signed graph.

Throws

An error if a suitable graph cannot be generated within the maximum number of iterations.

QuantumPrimer.classify_state_frustration_index — Function

classify_state_frustration_index(n::Int, threshold::Int, label::Int; max_iter::Int=10000, use_line_graph::Bool=false) -> SignedGraph

Classifies a random signed graph based on its frustration index.

Arguments

n::Int: The number of vertices.

threshold::Int: The threshold for the frustration index.

label::Int: The label indicating the desired classification (1 for frustration index above the threshold, -1 for frustration index below or equal to the threshold).

max_iter::Int: The maximum number of iterations to attempt, default is 10000.

use_line_graph::Bool: Whether to use the line graph of the random signed graph, default is false.

Returns

SignedGraph: The classified random signed graph.

Throws

An error if a suitable graph cannot be generated within the maximum number of iterations.

Frustration index

QuantumPrimer.find_cycles — Function

find_cycles(G::SignedGraph) -> Set{Tuple{Vector{Int}, Vector{Vector{Int}}}}

Finds all cycles in a given signed graph using depth-first search.

Arguments

G::SignedGraph: The input signed graph.

Returns

Set{Tuple{Vector{Int}, Vector{Vector{Int}}}}: A set of cycles, each represented as a tuple of vertex indices and edge paths.

QuantumPrimer.get_neighbors — Function

get_neighbors(G::SignedGraph, vertex::Int) -> Vector{Int}

Gets the neighbors of a given vertex in a signed graph.

Arguments

G::SignedGraph: The input signed graph.

vertex::Int: The vertex for which to find neighbors.

Returns

Vector{Int}: A vector of neighboring vertex indices.

QuantumPrimer.dfs_find_cycles! — Function

dfs_find_cycles!(G::SignedGraph, current_vertex::Int, start_vertex::Int, visited::BitVector, path::Vector{Int}, edge_path::Vector{Vector{Int}}, cycles::Set{Tuple{Vector{Int}, Vector{Vector{Int}}}})

Performs a depth-first search to find all cycles in a signed graph.

Arguments

G::SignedGraph: The input signed graph.

current_vertex::Int: The current vertex in the DFS.

start_vertex::Int: The starting vertex for the DFS.

visited::BitVector: A bit vector indicating visited vertices.

path::Vector{Int}: The current path in the DFS.

edge_path::Vector{Vector{Int}}: The current edge path in the DFS.

cycles::Set{Tuple{Vector{Int}, Vector{Vector{Int}}}}: The set of found cycles.

QuantumPrimer.evaluate_cycles_signs — Function

evaluate_cycles_signs(G::SignedGraph) -> Vector{Int}

Evaluates the product of signs for the cycles in a given signed graph.

Arguments

G::SignedGraph: The input signed graph.

Returns

Vector{Int}: A vector of signs for each cycle.

QuantumPrimer.count_unbalanced_cycles — Function

count_unbalanced_cycles(G::SignedGraph) -> Int

Counts the number of unbalanced cycles in a given signed graph.

Arguments

G::SignedGraph: The input signed graph.

Returns

Int: The number of unbalanced cycles.

QuantumPrimer.find_frustration_index — Function

find_frustration_index(G::SignedGraph) -> Tuple{Int, Vector{Int}}

Finds the frustration index of a given signed graph by changing the sign of the edges and checking the number of unbalanced cycles.

Arguments

G::SignedGraph: The input signed graph.

Returns

Tuple{Int, Vector{Int}}: The frustration index and the corresponding sign changes.

QuantumPrimer.balance_signed_graph — Function

balance_signed_graph(graph::SignedGraph) -> SignedGraph

Balances a signed graph by adjusting the signs of the edges to make the frustration index zero.

Arguments

graph::SignedGraph: The input signed graph.

Returns

SignedGraph: The balanced signed graph.

QuantumPrimer.reduce_frustration_signed_graph — Function

reduce_frustration_signed_graph(graph::SignedGraph, target::Int) -> SignedGraph

Reduces the frustration index of a signed graph by "flipping" a target number of -1 in the sign_changes vector.

Arguments

graph::SignedGraph: The input signed graph.

target::Int: The number of sign changes to reduce the frustration index.

Returns

SignedGraph: The signed graph with reduced frustration index.

Utility graph function

QuantumPrimer.create_random_signed_graph — Function

create_random_signed_graph(n::Int) -> SignedGraph

Creates a random signed graph with n vertices.

Arguments

n::Int: The number of vertices.

Returns

SignedGraph: The resulting random signed graph.

QuantumPrimer.plot_signed_graph — Function

plot_signed_graph(graph::SignedGraph) -> Tuple{SimpleGraph, Vector{Tuple{Float64, Float64}}, Matrix{Symbol}}

Plots a signed graph by creating a simple graph representation, a dataset of vertex positions, and an edge color matrix.

Arguments

graph::SignedGraph: The input signed graph.

Returns

Tuple{SimpleGraph, Vector{Tuple{Float64, Float64}}, Matrix{Symbol}}: A tuple containing the simple graph, the dataset of vertex positions, and the edge color matrix.