Loss documentation

Loss evaluation

QuantumPrimer.eval_loss — Function

eval_loss(state::ArrayReg, model::AbstractModel, cost::GeneralCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a given quantum state and model using a general cost function.

Arguments

state::ArrayReg: The quantum state.

model::AbstractModel: The quantum model.

cost::GeneralCost: The general cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(state::ArrayReg, model::AbstractModel, cost::CircuitCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a given quantum state and model using a circuit cost function.

Arguments

state::ArrayReg: The quantum state.

model::AbstractModel: The quantum model.

cost::CircuitCost: The circuit cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(states::NTuple{2, ArrayReg}, models::NTuple{2, AbstractModel}, cost::CircuitCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a pair of quantum states and models using a circuit cost function.

Arguments

states::NTuple{2, ArrayReg}: The tuple of quantum states.

models::NTuple{2, AbstractModel}: The tuple of quantum models.

cost::CircuitCost: The circuit cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(states::NTuple{2, ArrayReg}, models::NTuple{<:Any, AbstractModel}, cost::CircuitCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a pair of quantum states and models using a circuit cost function.

Arguments

states::NTuple{2, ArrayReg}: The tuple of quantum states.

models::NTuple{<:Any, AbstractModel}: The tuple of quantum models.

cost::CircuitCost: The circuit cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(states::NTuple{2, ArrayReg}, model::AbstractModel, cost::GeneralCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a pair of quantum states and a model using a general cost function.

Arguments

states::NTuple{2, ArrayReg}: The tuple of quantum states.

model::AbstractModel: The quantum model.

cost::GeneralCost: The general cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(states::NTuple{2, ArrayReg}, model::AbstractModel, cost::CircuitCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a pair of quantum states and a model using a general cost function.

Arguments

states::NTuple{2, ArrayReg}: The tuple of quantum states.

model::AbstractModel: The quantum model.

cost::CircuitCost: The circuit cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(state::ArrayReg, models::NTuple{2, AbstractModel}, cost::GeneralCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a quantum state and a tuple of models using a general cost function.

Arguments

state::ArrayReg: The quantum state.

models::NTuple{2, AbstractModel}: The tuple of quantum models.

cost::GeneralCost: The general cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

eval_loss(state::ArrayReg, models::NTuple{2, AbstractModel}, cost::CircuitCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the loss for a quantum state and a tuple of models using a circuit cost function.

Arguments

state::ArrayReg: The quantum state.

models::NTuple{2, AbstractModel}: The tuple of quantum models.

cost::CircuitCost: The circuit cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated loss.

QuantumPrimer.eval_full_loss — Function

eval_full_loss(data::AbstractData, model::Union{AbstractModel, NTuple{<:Any, AbstractModel}}, cost::AbstractCost; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Evaluates the full loss over a dataset for a given model or tuple of models.

Arguments

data::AbstractData: The dataset containing quantum states and corresponding labels.

model::Union{AbstractModel, NTuple{<:Any, AbstractModel}}: The quantum model or tuple of models.

cost::AbstractCost: The cost function.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The evaluated full loss.

Loss regularization

QuantumPrimer.regularize_loss — Function

regularize_loss(loss::Float64, model::AbstractModel; lambda=1.0::Float64, regularization=:nothing::Symbol) -> Float64

Regularizes the loss using L1 or L2 regularization.

Arguments

loss::Float64: The initial loss value.

model::AbstractModel: The quantum model.

lambda::Float64: The regularization parameter, default is 1.0.

regularization::Symbol: The type of regularization (:l1 or :l2), default is :nothing.

Returns

Float64: The regularized loss.