Model documentation

Structure

QuantumPrimer.AbstractModel — Type

AbstractModel

An abstract type for representing quantum models.

QuantumPrimer.GeneralModel — Type

GeneralModel{NN<:Int, CC<:ChainBlock, AA<:Function, TT<:AbstractVector{T} where T<:Real}

A mutable struct representing a general quantum model.

Fields

n::NN: The number of qubits.

n_layers::NN: The number of layers in the model, default is ceil(Int, log2(n)).

circ::CC: The quantum circuit.

ansatz::AA: The ansatz function.

params::TT: The parameters of the model, default is an empty vector.

Constructor

GeneralModel(n::NN, circ::CC, ansatz::AA; n_layers::NN=ceil(Int, log2(n)), params::TT=Float64[]): Creates a GeneralModel instance.

QuantumPrimer.InvariantModel — Type

InvariantModel{NN<:Int, CC<:ChainBlock, AA<:Function, TT<:AbstractVector{T} where T<:Real}

A mutable struct representing an invariant quantum model.

Fields

n::NN: The number of qubits.

n_layers::NN: The number of layers in the model, default is ceil(Int, log2(n)).

circ::CC: The quantum circuit.

ansatz::AA: The ansatz function.

params::TT: The parameters of the model, default is an empty vector.

Constructor

InvariantModel(n::NN, circ::CC, ansatz::AA; n_layers::NN=ceil(Int, log2(n)), params::TT=Float64[]): Creates an InvariantModel instance.

Parameter functions

QuantumPrimer.initialize_params — Function

initialize_params(model::GeneralModel)

Initializes the parameters of a GeneralModel instance.

Arguments

model::GeneralModel: The general quantum model.

Throws

An error if the number of qubits is not a power of 2.

initialize_params(model::InvariantModel)

Initializes the parameters of an InvariantModel instance.

Arguments

model::InvariantModel: The invariant quantum model.

Throws

An error if the number of qubits is not a power of 2.

QuantumPrimer.expand_params — Function

expand_params(model::GeneralModel) -> AbstractVector{T} where T<:Real

Expands the parameters of a GeneralModel instance.

Arguments

model::GeneralModel: The general quantum model.

Returns

AbstractVector{T} where T<:Real: The expanded parameters.

expand_params(model::InvariantModel) -> AbstractVector{T} where T<:Real

Expands the parameters of an InvariantModel instance.

Arguments

model::InvariantModel: The invariant quantum model.

Returns

AbstractVector{T} where T<:Real: The expanded parameters.

QuantumPrimer.reduce_params — Function

reduce_params(model::GeneralModel, vect::AbstractVector{T} where T<:Real) -> AbstractVector{T} where T<:Real

Reduces the parameters of a GeneralModel instance.

Arguments

model::GeneralModel: The general quantum model.

vect::AbstractVector{T} where T<:Real: The vector of parameters to be reduced.

Returns

AbstractVector{T} where T<:Real: The reduced parameters.

reduce_params(model::InvariantModel, vect::AbstractVector{T} where T<:Real) -> AbstractVector{T} where T<:Real

Reduces the parameters of an InvariantModel instance.

Arguments

model::InvariantModel: The invariant quantum model.

vect::AbstractVector{T} where T<:Real: The vector of parameters to be reduced.

Returns

AbstractVector{T} where T<:Real: The reduced parameters.