QCNN documentation

Data

QComp.Data — Type

struct Data{S <: Vector{SS} where SS <: ArrayReg, L <: Vector{LL} where LL <: Real}

Structure representing input data for the model.

Fields

s::S: Input states.

l::L: Labels.

Gradient

QComp.eval_grad — Function

eval_grad(state, p::GenericParams)

Compute the gradient of the cost function for a given state.

Arguments

state: Input state.

p::GenericParams: Model parameters.

Returns

The gradient of the cost function.

eval_grad(state, p::InvariantParams)

Compute the gradient of the cost function for a given state with parameters with translational invariance.

Arguments

state: Input state.

p::InvariantParams: Model parameters.

Returns

The gradient of the cost function.

QComp.eval_full_grad — Function

eval_full_grad(d::Data, p::InvariantParams, sig)

Compute the full gradient of the cost function for input data with parameters with translational invariance.

Arguments

d::Data: Input data and labels.

p::InvariantParams: Model parameters.

sig: Boolean indicating whether to use sigmoid activation.

Returns

The total gradient of the cost function.

Loss

QComp.sigmoid — Function

sigmoid(x)

Compute the sigmoid function.

Arguments

x: Input value.

Returns

The sigmoid of the input value.

QComp.eval_loss — Function

eval_loss(state, p::GenericParams)

Evaluate the loss of a model (with no invariance) applied to a state.

Arguments

state: Input state.

p::GenericParams: Model parameters.

Returns

The loss value.

eval_loss(state, p::InvariantParams)

Evaluate the loss of a translational invariant model applied to a state.

Arguments

state: Input state.

p::InvariantParams: Model parameters.

Returns

The loss value.

QComp.eval_full_loss — Function

eval_full_loss(d::Data, p::Params, sig)

Evaluate the full mean squared error (MSE) loss of a model on input data.

Arguments

d::Data: Input data and labels.

p::Params: Model parameters.

sig: Boolean indicating whether to use sigmoid activation.

Returns

The total mean squared error loss.

Parameters

QComp.Params — Type

abstract type Params end

Abstract type representing parameters for the QCNN.

QComp.GenericParams — Type

Base.@kwdef mutable struct GenericParams{NN<:Integer, CC<:ChainBlock, TT<:AbstractVector{T} where T<:Real} <:Params

Defines parameters for a quantum circuit without translational invariance.

Fields

n::NN: Dimension of the quantum register (must be a power of 2).

circ::CC: Circuit structure of the QCNN.

params::TT = Float64[]: Vector of parameters, initialized to an empty vector of Float64 values.

QComp.InvariantParams — Type

Base.@kwdef mutable struct InvariantParams{NN<:Integer, CC<:ChainBlock, TT<:AbstractVector{T} where T<:Real} <:Params

Defines parameters for a quantum circuit with translational invariance.

Fields

n::NN: Dimension of the quantum register (must be a power of 2).

circ::CC: Circuit structure of the QCNN.

params::TT = Float64[]: Vector of parameters, initialized to an empty vector of Float64 values.

QComp.initialize_params — Function

initialize_params(p::InvariantParams)

Initialize parameters for a quantum circuit with translational invariance.

Arguments

p::InvariantParams: Parameters object containing circuit information.

Errors

Throws an error if the register dimension is not a power of 2.

QComp.expand_params — Function

expand_params(p::InvariantParams)

Expand unique translational invariant parameters to a full vector.

Arguments

p::InvariantParams: Parameters object containing circuit information.

Returns

A vector representing expanded parameters.

QComp.reduce_params — Function

reduce_params(n, params)

Reduce a vector of parameters to the unique set of parameters in the translational invariant QCNN. Performs the sum of the corresponding parameters in the circuit. Used in evalfullgrad.

Arguments

n: Dimension of the quantum register.

params: Vector of "extended" parameters.

Returns

A vector of unique parameters.

QCNN

QComp.conv_Ry — Function

conv_Ry(n, i, j)

Ansatz for the convolutional layer. Use Ry gate to keep ansatz real (adapted from arxiv:2108.00661v2).

Arguments

n: Dimension of the quantum register.

i: Index of the first qubit.

j: Index of the second qubit.

Returns

A quantum circuit implementing a simple convolutional layer with Ry gates.

QComp.conv_Ry2 — Function

conv_Ry2(n, i, j)

Ansatz for the convolutional layer. Use Ry gate to keep ansatz real (adapted from arxiv:2108.00661v2).

Arguments

n: Dimension of the quantum register.

i: Index of the first qubit.

j: Index of the second qubit.

Returns

A quantum circuit implementing a convolutional layer with Ry gates.

QComp.conv_SU4 — Function

conv_SU4(n, i, j)

Ansatz for the convolutional layer, applying an arbitrary SU(4) gate (adapted from arxiv:2108.00661v2).

Arguments

n: Dimension of the quantum register.

i: Index of the first qubit.

j: Index of the second qubit.

Returns

A quantum circuit implementing a convolutional layer.

QComp.build_QCNN — Function

build_QCNN(n;anz=conv_Ry)

Build a quantum circuit for a Quantum Convolutional Neural Network (QCNN).

Arguments

n: Dimension of the quantum register.

anz: Convolutional ansatz

Returns

A quantum circuit representing the QCNN.

TrainTest

QComp.test_model — Function

test_model(d::Data, p::Params, sig)

Evaluate classification accuracy of the model.

Arguments

d::Data: Input data and labels.

p::Params: Model parameters.

sig: Boolean indicating whether to use sigmoid activation.

Returns

preds: Predicted values.

suc_rate: Success rate.

suc_inds: Indices of successful predictions.

QComp.train_test_model — Function

train_test_model(d1::Data, d2::Data, p::Params, iters, sig, lr; output=false)

Train and test the model, outputting final predictions and train/test accuracy.

Arguments

d1::Data: Training data and labels.

d2::Data: Testing data and labels.

p::Params: Initial model parameters.

iters: Number of training iterations.

sig: Boolean indicating whether to use sigmoid activation.

lr: Learning rate for optimization.

output: Boolean indicating whether to output intermediate training information.

Returns

p.params: Final model parameters.

loss_track: Track of loss values during training.

tr_track: Track of training accuracy during training.

te_track: Track of testing accuracy during training.

tr_preds: Predictions on training data.

te_preds: Predictions on testing data.