QSP documentation

QuantumPrimer.W — Function

W(a::Float64) -> ChainBlock

Creates a quantum circuit block for the W gate.

Arguments

a::Float64: The parameter for the W gate.

Returns

ChainBlock: The quantum circuit block for the W gate.

QuantumPrimer.S — Function

S(phi::Float64) -> Rz

Creates a quantum gate for the S gate.

Arguments

phi::Float64: The parameter for the S gate.

Returns

Rz: The quantum gate for the S gate.

QuantumPrimer.Usp — Function

Usp(phis::Vector{Float64}, a::Float64) -> ChainBlock

Creates a quantum circuit for the Usp unitary.

Arguments

phis::Vector{Float64}: The vector of phase parameters.

a::Float64: The parameter for the W gate.

Returns

ChainBlock: The quantum circuit for the Usp unitary.

QuantumPrimer.eval_Usp — Function

eval_Usp(x::Float64, phis::Vector{Float64}) -> Float64

Evaluates the Usp unitary for a given input and phase parameters.

Arguments

x::Float64: The input parameter.

phis::Vector{Float64}: The vector of phase parameters.

Returns

Float64: The evaluated value of the Usp unitary.

QuantumPrimer.loss — Function

loss(target::Function, xs::Vector{Float64}, phis::Vector{Float64}) -> Float64

Computes the loss function for a given target function, inputs, and phase parameters.

Arguments

target::Function: The target function.

xs::Vector{Float64}: The vector of input parameters.

phis::Vector{Float64}: The vector of phase parameters.

Returns

Float64: The computed loss.

QuantumPrimer.pcp — Function

pcp(n::Int, phi::Float64) -> ChainBlock

Creates a projected controlled phase gate.

Arguments

n::Int: The number of qubits.

phi::Float64: The phase parameter.

Returns

ChainBlock: The quantum circuit block for the projected controlled phase gate.

QuantumPrimer.block_encode2 — Function

block_encode2(n::Int, A::AbstractMatrix) -> ChainBlock

Creates a block encoding for a given matrix.

Arguments

n::Int: The number of qubits.

A::AbstractMatrix: The matrix to be block encoded.

Returns

ChainBlock: The quantum circuit block for the block encoding.

QuantumPrimer.QSVT_square — Function

QSVT_square(n::Int, d::Int, phis::Vector{Float64}, A::AbstractMatrix) -> ChainBlock

Creates a quantum singular value transformation (QSVT) circuit for a given matrix.

Arguments

n::Int: The number of qubits.

d::Int: The degree of the polynomial.

phis::Vector{Float64}: The vector of phase parameters.

A::AbstractMatrix: The matrix to be transformed.

Returns

ChainBlock: The quantum circuit for the QSVT.