Utility documentation

QuantumPrimer.haar_random_unitary — Function

haar_random_unitary(n::Int) -> Matrix{ComplexF64}

Generates a Haar-random unitary matrix of size n x n.

Arguments

n::Int: The size of the unitary matrix.

Returns

Matrix{ComplexF64}: The Haar-random unitary matrix.

QuantumPrimer.wasserstein_distance — Function

wasserstein_distance(p::Vector{Float64}, q::Vector{Float64}) -> Float64

Computes the Wasserstein distance between two probability distributions.

Arguments

p::Vector{Float64}: The first probability distribution.

q::Vector{Float64}: The second probability distribution.

Returns

Float64: The Wasserstein distance.

Throws

An error if the input vectors are not valid probability distributions.

QuantumPrimer.KL_divergence — Function

KL_divergence(p::Vector{Float64}, q::Vector{Float64}) -> Float64

Computes the Kullback-Leibler (KL) divergence between two probability distributions.

Arguments

p::Vector{Float64}: The first probability distribution.

q::Vector{Float64}: The second probability distribution.

Returns

Float64: The KL divergence.

Throws

An error if the input vectors are not valid probability distributions or if p and q have different lengths.

An error if p is non-zero where q is zero.

QuantumPrimer.density_matrix_from_vector — Function

density_matrix_from_vector(states::Vector{Vector{Int}}; coeffs=nothing::Union{Nothing, Vector{Float64}})

Create a density matrix from a vector of integer vectors. Each integer vector is converted to an ArrayReg using bit_literal.

Arguments

states::Vector{Vector{Int}}: A vector of integer vectors to be included in the density matrix.

coeffs::Union{Nothing, Vector{Float64}}: Optional coefficients for the states. If provided, the coefficients are used in the density matrix calculation.

Returns

A density matrix created from the given integer vectors and coefficients.

QuantumPrimer.register_from_vector — Function

register_from_vector(states::Vector{Vector{Int}}; coeffs=nothing::Union{Nothing, Vector{Float64}})

Create a quantum register from a vector of integer vectors. Each integer vector is converted to an ArrayReg using bit_literal.

Arguments

states::Vector{Vector{Int}}: A vector of integer vectors to be included in the register.

coeffs::Union{Nothing, Vector{Float64}}: Optional coefficients for the states. If provided, the coefficients are normalized.

Returns

A quantum register created from the given integer vectors and coefficients.

QuantumPrimer.pauli_decomposition — Function

pauli_decomposition(matrix::Union{ChainBlock, Matrix{ComplexF64}, Matrix{T}} where T<:Real)

Decompose a matrix into a dictionary of Pauli strings.

Arguments

matrix::Union{ChainBlock, Matrix{ComplexF64}, Matrix{T}} where T<:Real: The input matrix to be decomposed. The matrix must be square with dimensions 2^n x 2^n.

Returns

A dictionary where the keys are vectors of integers representing Pauli strings, and the values are the corresponding coefficients.

QuantumPrimer.pauli_decomposition_kronecker — Function

pauli_decomposition_kronecker(matrix::Union{ChainBlock, Matrix{ComplexF64}, Matrix{T}} where T<:Real)

Decompose a matrix into a dictionary of Pauli strings using the package Kronecker.

Arguments

matrix::Union{ChainBlock, Matrix{ComplexF64}, Matrix{T}} where T<:Real: The input matrix to be decomposed. The matrix must be square with dimensions 2^n x 2^n.

Returns

A dictionary where the keys are vectors of integers representing Pauli strings, and the values are the corresponding coefficients.

QuantumPrimer.save_circuit_to_txt — Function

save_circuit_to_txt(circuit::AbstractBlock; filename::String="output.txt")

Save a quantum circuit to a text file as a console output.

Arguments

circuit::AbstractBlock: The quantum circuit to be saved.

filename::String="yao_circuit.txt": The name of the file to save the circuit to. Defaults to "yao_circuit.txt".

Returns

Nothing. The function writes the circuit to the specified file.

QuantumPrimer.normalize_vector — Function

normalize_vector(vector::Vector{T} where T <: Number) -> Vector{T}

Normalize a vector by dividing each element by the vector's Euclidean norm.

Arguments

vector::Vector{T} where T <: Number: The input vector to be normalized.

Returns

Vector{T}: The normalized vector.

Example

```julia v = [3.0, 4.0] normalize_vector(v) # returns [0.6, 0.8]

QuantumPrimer.prepare_dos_encoded_state — Function

prepare_dos_encoded_state(unitary::Union{AbstractMatrix, AbstractBlock}, n_times::Int; nshots::Int=1000, extended_output::Bool=false, focussed_output::Bool=true)

Prepare a state for the DOS algorithm by encoding a unitary matrix into a quantum state.

Arguments

unitary::Union{AbstractMatrix, AbstractBlock}: The unitary matrix to be encoded.

n_times::Int: The number of times to apply the unitary matrix.

nshots::Int=1000: The number of shots to use for the measurement.

extended_output::Bool=false: If true, the function returns the state and the measurement counts. If false, the function returns only the state.

focussed_output::Bool=true: If true, the function returns the state with the ancilla qubits removed. If false, the function returns the full state.

Returns

If extended_output is true, the function returns a tuple (dos_state, counts_register_after_qpe). If extended_output is false, the function returns only dos_state.