Variational quantum eigensolver (VQE) documentation

QuantumPrimer.circ_vqe — Function

circ_vqe(H::AbstractMatrix; ansatz=circ_HEA::Function, depth::Union{Symbol, Int}=:nothing, iters::Int=500, verbose::Bool=false)

Prepare a quantum circuit to approximate the ground state of the Hamiltonian H, using the Variational Quantum Eigensolver (VQE) approach.

Arguments

graph::Graph: The graph from which the modified Laplacian matrix is derived.

Keyword Arguments

depth::Union{Symbol, Int}=:nothing: The depth of the hardware-efficient ansatz (HEA) used in the circuit. If set to :nothing, the depth is automatically determined based on the logarithm of the number of nodes in the graph.

iters::Int=500: The number of iterations for the VQE optimization process.

verbose::Bool=false: If true, prints the energy distance at every 50th iteration and at the beginning.

Returns

QuantumCircuit: The optimized quantum circuit that approximates the ground state of the modified Laplacian matrix.