Circuits documentation

Basic circuits

QuantumPrimer.circ_gate_single — Function

circ_gate_single(n::Int, i::Int, gate::ConstantGate) -> ChainBlock

Creates a quantum circuit with a single gate applied to the i-th qubit.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the qubit to which the gate is applied.

gate::ConstantGate: The gate to be applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_gate_n — Function

circ_gate_n(n::Int, gate::ConstantGate) -> ChainBlock

Creates a quantum circuit with the same gate applied to all qubits.

Arguments

n::Int: The total number of qubits.

gate::ConstantGate: The gate to be applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_gate_sum — Function

circ_gate_sum(n::Int, gate::ConstantGate) -> ChainBlock

Creates a quantum circuit with the sum of the same gate applied to all qubits.

Arguments

n::Int: The total number of qubits.

gate::ConstantGate: The gate to be applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_gate_where — Function

circ_gate_where(n::Int, gate::ConstantGate, ones_where::Vector{Int}) -> ChainBlock

Creates a quantum circuit with the gate applied to specific qubits.

Arguments

n::Int: The total number of qubits.

gate::ConstantGate: The gate to be applied.

ones_where::Vector{Int}: The indices of the qubits to which the gate is applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_block_single — Function

circ_block_single(n::Int, i::Int, block::ChainBlock{2}) -> ChainBlock

Creates a quantum circuit with a single block applied to the i-th qubit.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the qubit to which the block is applied.

block::ChainBlock{2}: The block to be applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_block_n — Function

circ_block_n(n::Int, block::ChainBlock{2}) -> ChainBlock

Creates a quantum circuit with the same block applied to all qubits.

Arguments

n::Int: The total number of qubits.

block::ChainBlock{2}: The block to be applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_block_sum — Function

circ_block_sum(n::Int, block::ChainBlock{2}) -> ChainBlock

Creates a quantum circuit with the sum of the same block applied to all qubits.

Arguments

n::Int: The total number of qubits.

block::ChainBlock{2}: The block to be applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_block_where — Function

circ_block_where(n::Int, block::ChainBlock{2}, ones_where::Vector{Int}) -> ChainBlock

Creates a quantum circuit with the block applied to specific qubits.

Arguments

n::Int: The total number of qubits.

block::ChainBlock{2}: The block to be applied.

ones_where::Vector{Int}: The indices of the qubits to which the block is applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_X — Function

circ_X(n::Int, i::Int) -> ChainBlock

Creates a quantum circuit with an X gate applied to the i-th qubit.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the qubit to which the X gate is applied.

Returns

ChainBlock: The resulting quantum circuit.

circ_X(n::Int) -> ChainBlock

Creates a quantum circuit with an X gate applied to the first qubit.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Y — Function

circ_Y(n::Int, i::Int) -> ChainBlock

Creates a quantum circuit with a Y gate applied to the i-th qubit.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the qubit to which the Y gate is applied.

Returns

ChainBlock: The resulting quantum circuit.

circ_Y(n::Int) -> ChainBlock

Creates a quantum circuit with a Y gate applied to the first qubit.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Z — Function

circ_Z(n::Int, i::Int) -> ChainBlock

Creates a quantum circuit with a Z gate applied to the i-th qubit.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the qubit to which the Z gate is applied.

Returns

ChainBlock: The resulting quantum circuit.

circ_Z(n::Int) -> ChainBlock

Creates a quantum circuit with a Z gate applied to the first qubit.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_X_where — Function

circ_X_where(n::Int, ones_where::Vector{Int}) -> ChainBlock

Creates a quantum circuit with X gates applied to specific qubits.

Arguments

n::Int: The total number of qubits.

ones_where::Vector{Int}: The indices of the qubits to which the X gates are applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Y_where — Function

circ_Y_where(n::Int, ones_where::Vector{Int}) -> ChainBlock

Creates a quantum circuit with Y gates applied to specific qubits.

Arguments

n::Int: The total number of qubits.

ones_where::Vector{Int}: The indices of the qubits to which the Y gates are applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Z_where — Function

circ_Z_where(n::Int, ones_where::Vector{Int}) -> ChainBlock

Creates a quantum circuit with Z gates applied to specific qubits.

Arguments

n::Int: The total number of qubits.

ones_where::Vector{Int}: The indices of the qubits to which the Z gates are applied.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Xn — Function

circ_Xn(n::Int) -> ChainBlock

Creates a quantum circuit with X gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Yn — Function

circ_Yn(n::Int) -> ChainBlock

Creates a quantum circuit with Y gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Zn — Function

circ_Zn(n::Int) -> ChainBlock

Creates a quantum circuit with Z gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Xsum — Function

circ_Xsum(n::Int) -> ChainBlock

Creates a quantum circuit with the sum of X gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Ysum — Function

circ_Ysum(n::Int) -> ChainBlock

Creates a quantum circuit with the sum of Y gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Zsum — Function

circ_Zsum(n::Int) -> ChainBlock

Creates a quantum circuit with the sum of Z gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Hn — Function

circ_Hn(n::Int) -> ChainBlock

Creates a quantum circuit with Hadamard gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_CNOT_2regs — Function

circ_CNOT_2regs(n::Int) -> ChainBlock

Creates a quantum circuit with CNOT gates applied between two registers.

Arguments

n::Int: The total number of qubits in each register.

Returns

ChainBlock: The resulting quantum circuit.

Convolutional circuits

QuantumPrimer.circ_Ry_simple_conv — Function

circ_Ry_simple_conv(n::Int, i::Int, j::Int) -> ChainBlock

Creates a simple quantum circuit with Ry gates and a controlled-X gate.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the first qubit.

j::Int: The index of the second qubit.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_Ry_conv — Function

circ_Ry_conv(n::Int, i::Int, j::Int) -> ChainBlock

Creates a quantum circuit with multiple Ry gates and controlled-X gates.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the first qubit.

j::Int: The index of the second qubit.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_SU4_conv — Function

circ_SU4_conv(n::Int, i::Int, j::Int) -> ChainBlock

Creates a quantum circuit with SU(4) gates and controlled-X gates.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the first qubit.

j::Int: The index of the second qubit.

Returns

ChainBlock: The resulting quantum circuit.

Layer circuits

QuantumPrimer.circ_Rx_layer — Function

circ_Rx_layer(n::Int) -> ChainBlock

Creates a quantum circuit layer with Rx gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit layer.

QuantumPrimer.circ_Ry_layer — Function

circ_Ry_layer(n::Int) -> ChainBlock

Creates a quantum circuit layer with Ry gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit layer.

QuantumPrimer.circ_Rz_layer — Function

circ_Rz_layer(n::Int) -> ChainBlock

Creates a quantum circuit layer with Rz gates applied to all qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit layer.

QuantumPrimer.circ_CNOT_layer — Function

circ_CNOT_layer(n::Int) -> ChainBlock

Creates a quantum circuit layer with CNOT gates applied between adjacent qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit layer.

QuantumPrimer.circ_CZ_layer — Function

circ_CZ_layer(n::Int) -> ChainBlock

Creates a quantum circuit layer with CZ gates applied between adjacent qubits.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit layer.

QNN circuits

QuantumPrimer.circ_HEA — Function

circ_HEA(n::Int) -> ChainBlock

Creates a Hardware Efficient Ansatz (HEA) quantum circuit.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_HEA_real — Function

circ_HEA_real(n::Int) -> ChainBlock

Creates a Hardware Efficient Ansatz (HEA) quantum circuit. This version is for real-valued ground states.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

Hypergraph states circuits

QuantumPrimer.circ_phase_flip — Function

circ_phase_flip(n::Int, x::Int) -> ChainBlock

Creates a quantum circuit that performs a phase flip on the state x.

Arguments

n::Int: The total number of qubits.

x::Int: The state to be phase-flipped.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_hypergraph_state — Function

circ_hypergraph_state(vec::Vector{Int}) -> ChainBlock

Creates a quantum circuit that prepares a hypergraph state.

Arguments

vec::Vector{Int}: The vector representing the hypergraph state.

Returns

ChainBlock: The resulting quantum circuit.

circ_hypergraph_state(n::Int, edges::Vector{Vector{Int}}) -> ChainBlock

Creates a quantum circuit that prepares a hypergraph state from a list of edges.

Arguments

n::Int: The number of qubits in the circuit.

edges::Vector{Vector{Int}}: A vector of vectors, where each inner vector represents an (hyper-)edge in the hypergraph.

Returns

ChainBlock: The resulting quantum circuit.

Measurement circuits

QuantumPrimer.circ_swap_decomposed — Function

circ_swap_decomposed(n::Int, i::Int, j::Int) -> ChainBlock

Creates a quantum circuit that performs a decomposed SWAP operation between qubits i and j.

Arguments

n::Int: The total number of qubits.

i::Int: The index of the first qubit.

j::Int: The index of the second qubit.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_swap_all — Function

circ_swap_all(n::Int; decompose::Bool=false) -> ChainBlock

Creates a quantum circuit that performs SWAP operations between pairs of qubits.

Arguments

n::Int: The total number of qubits.

decompose::Bool: Whether to use the decomposed SWAP operation.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_swap_test — Function

circ_swap_test(n::Int) -> ChainBlock

Creates a quantum circuit that performs a SWAP test.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_destructive_swap_test — Function

circ_destructive_swap_test(n::Int) -> ChainBlock

Creates a quantum circuit that performs a destructive SWAP test.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_obs_times_swap — Function

circ_obs_times_swap(n::Int, obs::Union{ChainBlock, Add}) -> ChainBlock

Creates a quantum circuit that performs an observable times SWAP operation.

Arguments

n::Int: The total number of qubits.

obs::Union{ChainBlock, Add}: The observable to be applied.

Returns

ChainBlock: The resulting quantum circuit.

LCU

QuantumPrimer.circ_LCU — Function

circ_LCU(n::Int, U_vec::Vector{ChainBlock}; initial_layer::Bool=true, final_layer::Bool=true) -> ChainBlock

Creates a quantum circuit that performs a Linear Combination of Unitaries (LCU) operation.

Arguments

n::Int: The total number of qubits.

U_vec::Vector{ChainBlock}: The vector of unitary operations.

initial_layer::Bool: Whether to include the initial layer of Hadamard gates.

final_layer::Bool: Whether to include the final layer of Hadamard gates.

Returns

ChainBlock: The resulting quantum circuit.

Haar random unitary

QuantumPrimer.circ_haar_random_unitary — Function

circ_haar_random_unitary(n::Int) -> ChainBlock

Generates a quantum circuit with a Haar-random unitary matrix of size 2^n x 2^n.

Arguments

n::Int: The number of qubits.

Returns

ChainBlock: The quantum circuit with the Haar-random unitary matrix.

QPE circuits

QuantumPrimer.circ_phase_encoding — Function

circ_phase_encoding(n::Int, unitary_block::AbstractBlock; reverse_qubits::Bool=false) -> ChainBlock

Creates a quantum circuit for phase encoding using a unitary block.

Arguments

n::Int: The number of qubits for the time register.

unitary_block::AbstractBlock: The unitary block to be applied.

reverse_qubits::Bool: Whether to reverse the order of qubits.

Returns

ChainBlock: The resulting quantum circuit.

circ_phase_encoding(n::Int, unitary_matrix::AbstractMatrix; reverse_qubits::Bool=false) -> ChainBlock

Creates a quantum circuit for phase encoding using a unitary matrix.

Arguments

n::Int: The number of qubits for the time register.

unitary_matrix::AbstractMatrix: The unitary matrix to be applied.

reverse_qubits::Bool: Whether to reverse the order of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_qft — Function

circ_qft(n::Int; reverse_qubits::Bool=false) -> ChainBlock

Creates a quantum circuit for the Quantum Fourier Transform (QFT).

Arguments

n::Int: The number of qubits.

reverse_qubits::Bool: Whether to reverse the order of qubits.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_qpe — Function

circ_qpe(n::Int, unitary_matrix::Union{AbstractMatrix, AbstractBlock}; reverse_qubits::Bool=false) -> ChainBlock

Creates a quantum circuit for Quantum Phase Estimation (QPE).

Arguments

n::Int: The number of qubits for the time register.

unitary_matrix::Union{AbstractMatrix, AbstractBlock}: The unitary matrix to be applied.

reverse_qubits::Bool: Whether to reverse the order of qubits.

Returns

ChainBlock: The resulting quantum circuit.

Utility circuits

QuantumPrimer.circ_purified_maximally_mixed — Function

circ_purified_maximally_mixed(n::Int) -> ChainBlock

Creates a quantum circuit that prepares a purified maximally mixed state.

Arguments

n::Int: The number of qubits in each register.

Returns

ChainBlock: The resulting quantum circuit.

QuantumPrimer.circ_append_ancillas — Function

circ_append_ancillas(circ::AbstractBlock, n_ancillas::Int; pos_ancillas::Symbol=:bottom) -> ChainBlock

Appends ancilla qubits to a quantum circuit.

Arguments

circ::AbstractBlock: The quantum circuit.

n_ancillas::Int: The number of ancilla qubits.

pos_ancillas::Symbol: The position of the ancilla qubits (:bottom or :top).

Returns

ChainBlock: The resulting quantum circuit with ancilla qubits.

QuantumPrimer.circ_reverse_order — Function

circ_reverse_order(n::Int) -> ChainBlock

Creates a quantum circuit that reverses the order of qubits using SWAP gates.

Arguments

n::Int: The total number of qubits.

Returns

ChainBlock: The resulting quantum circuit that reverses the order of the qubits.